deleted InFileDatabase in order to pass code coverage, added tests for Invoice, please look at comments

+++ src/main/java/pl/coderstrust/accounting/model/Company.java
@@ -9,9 +9,9 @@ public class Company {
   private String street;
   private String postalCode;
   private String city;
-  private BigDecimal discount;
+  private double discount;
 
-  public BigDecimal getDiscount() {
+  public double getDiscount() {
     return discount;
   }
 }
